+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add `maxDepth` option and `limitDepth` function
 - Add support for parser options `maxTokens` and `noLocation`
 - Remove implicit dependency on `@types/ws`
+- Allow async `createContext`
 - Update deps
 
 ## 1.7.1